requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.5.0
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==4.1.0
cachetools==5.3.2
celery==5.3.4
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
DateTime==5.2
Django==4.0.3
django-cors-headers==3.11.0
django-dotenv==1.4.2
django-filter==23.3
django-multiselectfield==0.1.12
django-smart-selects==1.6.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
et-xmlfile==1.1.0
frozenlist==1.4.0
google==3.0.0
google-api-core==2.15.0
google-auth==2.26.1
google-cloud-recaptcha-enterprise==1.16.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
idna==3.4
kombu==5.3.2
multidict==6.0.4
numpy==1.26.4
openpyxl==3.1.3
pandas==2.2.2
Pillow==9.5.0
prompt-toolkit==3.0.39
proto-plus==1.23.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
redis==5.0.1
requests==2.31.0
rsa==4.9
sib-api-v3-sdk==7.6.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.2
twilio==8.10.0
tzdata==2023.3
urllib3==2.0.7
vine==5.0.0
wcwidth==0.2.8
xlrd==2.0.1
XlsxWriter==3.2.0
yarl==1.9.2
zope.interface==6.0
